I removed the hardcoded (and BTW never checked) limit on 16 interfaces:
We do this by adding the "struct netbuf" to the "struct interface" because
- we avoid maintaining a separate array of pointers and malloc(3)ed struct's.
- we avoid therefore lots of malloc(3) and free(3).
- we need such a thing anyway for each interface, so why not put it in there
  (and libnet_ctx is also there).
- we do not longer neeed the interface-index as a separate field.

Does anyone know if the MTU of an interface may chnage (and if yes, when)?
Otherwise we could get rid of another field in that struct (and a few lines
of code).

I compile-tested with `make build_all USE_LIBNET=1` but didn't run it.

Downside: One must initialize this struct (or at least the pointer to the
          actual buffer). Therefore the one line change in src/unix/ifnet.c.
          Therefore I broke very probably the Windows port and I can' figure out
          if and where to add such an equivalent thing.

Other changes:
- The libnet_ctx field in struct interface is also in a
  #ifdef USE_LIBNET ... #endif section.