- added "-fomit-frame-pointer" to the normal build . start a new era on configuration: We introduce a new higher level configuratino layer to enable or disable features/configuration options:
Rationale: Take "debug build" as prime example.
To have a real debug build, you want (at least) "-g" in CFLAGS and no
stripping. It is cumbersome and error prone to edit that by hand (or pass
parameters to the "make") and with a new options somewhere (e.g.
"-fomit-frame-pointer" from above) one has to adjust that.
For these (so called) high leveloptions, I started with
---  snip  ----
# activate debugging with 1 or deactivate with 0
DEBUG ?= 0

# the optimize option to be set for gcc
OPTIMIZE ?= -O2

# enable mudflap with 1 or deactivate with 0
# you need a recent enough gcc and the libmudflap installed
MUDFLAP ?= 0
---  snip  ----
Feel free to propose other useful ones (and submit patches;-).

The lower part, which is basically the former Makefile.inc, gets now "ifeq"
etc. so that a simple "make DEBUG=1" makes a real debug build.
ANd this is available and usefule to the plugin Makeifle's too of course.