From Hannes Gredler <hannes@gredler.at>: refactor the timer implementation
-slash all timeout functions and the corresponding infrastructure
-slash all event functions and the corresponding infrastructure

  that means all polling based timer comparisions have been removed.
  the scheduler rate can be now brought down to 10ms without much
  additional overhead.

-introduce a new wheel based timer similar to the BSD kernel implementation
 described in http://www.olsr.org/docs/wucs-95-23.pdf

 there are three major calls

  olsr_start_timer()
  olsr_stop_timer()
  olsr_change_timer()

alternatively there is a one-stop shop that figures out on its own what
to do.

  olsr_set_timer()

this comes handy for avoiding a lot of copy and paste code updating
all the client code.

the new timer API supports periodical and singleshot timers.
once the timer expires a provided function along with a context pointer
is called back. After the singleshot timer fires the used memeory for
the timer is not freed but rather returned to a free memory pool for
avoiding malloc() churn.

the implementation is very efficient as all timer start/stop/change operations
are done in O(1) complexity. walking the timer slot in the main scheduler loop
has a complexity of O(N/K) where K is 256 which should give reasonable defaults
for thousands of timers in the system.

as a sideeffect all wallclock timer references have been removed. now only
one non-wallclock timer obtained by the times() call is maintained.

note that the olsr_secure plugin does keep its wallclock based timer.
it is left to the olsrd_secure plugin authors to convert this to a
non-wallclock based timers to avoid disaster events like when the system
clock changes.