Make the socket parser more general
The general goal is to get to a flexible central event loop to allow
decoupling of long time actions (e.g. a slow browser over a slow
connection) from the rest.

We extend the central select() loop to get an additional select() for
file descriptors where the data will be handled immediately.
We do this via an additional callback function (as to not disturb the
current ones).
The caller must also specify, if he wants to be called on a possible
read andor write on the "immediate" and/or the "pollrate" callback
function.
While we're at it, we add a "void *" which is passed back to the caller.
So the caller can actually register more than one instance on different
file descriptors and differentiate the instances via that pointer. E.g.
the httpinfo plugin is now able to really accept more than one client
connection at the same time.