Cleanup the signal handlers
Not really belonging to that theme but I stumbled over it and it annoyed me
since ages:
The signal handlers where quite large functions very probably using system-calls
which one better do not use in a signal handler. Read Stevens'  "Programming in
the Unix environment" for more details into signal handlers.
The worse for Win32: Apperentyl Win32 needs two extra variables to let one
thread break out of the select() loop and continue afterwards with the shutdown
in the signal handler.

The simple and clean solution is quite easy: We add a state variable which is
changed in the signal handlers and do the whole action afterwards.

And the usual coding style cleanup.

This has been tested on Linux and Win32.