Replace the nanosleep() with select()
And now to the more gory stuff. We replace the nanosleep() in
the central loop with a select() (which doesn't has any file
descriptors there yet).

We basically select() and recalculate the timeout until the
time is over. And we make sure that we select() at least once -
even if the timeout is "0" (or < "0").
This is to give the funtions chance to do something even in
high-load situations.
The olsr_scheduler_sleep() function is thus superflous.