Factor out the buffer code and use it elsewhere
We factor out the code of the auto-expanding buffer from the txtinfo plugin
into the core (and a few functions ....).
We use that for the config-file generation and while we are at it,
we clean it up.
And we use it in the httpinfo plugin.
This saves a few KB in the olsrd, httpinfo and txtinfo plugin.