Hi,
with this patch you can now select an ULA (RFC 4193 --> replacement of the
site-local addresses) for the OLSR meshing by setting Ip6AddrType to
"unique-local".

And it has now an auto mode, too. If you set Ip6AddrType to "auto", it
try to find a site-local first, then an unique-local, then a global. Good
as the default config, so that you can easy setup an IPv6 mesh.

Warning: The *BSD code is untested, because I don't have this OS here.