Remove the olsr-specific duplicated types
Namely this are olsr_bool and the integer types with defined widhts.
For the first, we use <stdbool.h> as defined in C99 or emulate it. For
the latter, we use <stdint.h> or emulate it.

Apart from editing src/olsr_cfg.h by hand, the rest iwas done with
----  snip  ----
find src lib \( -name '*.[chy]' -o -name 'oscan.lex' \) -print | \
    xargs perl -w -i -p -e '\
        s/olsr_(u?)(8|16|32)_t/$1int$2_t/g;\
        s/olsr_bool/bool/g;\
        s/OLSR_TRUE/true/g;\
        s/OLSR_FALSE/false/g;'
----  snip  ----
Beware: The line breaks are there for readability. Just put the whole
on one line in your $SHELL.

If you have larger changes, the easiest forward proitng is probably:
- make a patch with your changes against the previous version (and
  save it away;-)
- `hg revert -a`
- `hg fetch`
- `perl -w -p -e '\
        s/olsr_(u?)(8|16|32)_t/$1int$2_t/g;\
        s/olsr_bool/bool/g;\
        s/OLSR_TRUE/true/g;\
        s/OLSR_FALSE/false/g;' < $YOUR_PATCH > $YOUR_PATCH-new`
- and apply your patch again.
For hg/mq people:
- `hg qpo -a`
- `hg fetch`
- `perl -w -i -p -e '\
        s/olsr_(u?)(8|16|32)_t/$1int$2_t/g;\
        s/olsr_bool/bool/g;\
        s/OLSR_TRUE/true/g;\
        s/OLSR_FALSE/false/g;' .hg/patches/*`
   Take care that you do not change the names in .hg/patches/series.