The rt_policy switch is somewhat superfluous because under Linux u always set policy routes. I left the setting in here as a convenient way to say "do not fiddle with unexpected things" with a single rt_policy=no
Besides that: rules setting is useful for smartgw/NIIT
stuff only. For standard policy-settings as they are
used with previous versions of olsrd they are not
required nor helpful. Also: I missed "rt_table_default=0"
which basically says: always use same table as rt_table.
If that is other than 254, please maintain the olsrd-default
route in table "default" as fallback.