Linux: correctly set plugin library soname
By convention, non-core libraries should start with the 'lib'
prefix on Linux systems. Currently this is not the case for the
olsrd plugin libraries.

Since renaming the libraries would break all configurations,
the soname of an olsrd plugin library is now correctly set
with the 'lib' prefix so that an 'ldconfig -l <library>' will
create the correct link.

This patch fixes the issue for Linux only, other platforms
are not touched since I don't have those platforms available.

To completely fix the issue, PLUGIN_FULLNAME must be set as:
  PLUGIN_FULLNAME ?= lib$(PLUGIN_NAME).so.$(PLUGIN_VER)
Doing this will rename all olsrd plugin libraries.

Signed-off-by: Ferry Huberts <f.huberts@mindef.nl>