smart gateway: add threshold configuration parameter
We can now keep choosing a better gateway by setting the
configuration setting "SmartGatewayThreshold" to a value
larger than zero.

The costs for a new smart gateway must be less than the
costs of the current smart gateway multiplied by
SmartGatewayThreshold in order for it to be chosen as
the new smart gateway.

This new configuration parameter has a similar function
and meaning as the NAT threshold configuration parameter.

Signed-off-by: Ferry Huberts <f.huberts@mindef.nl>